requirements.txt
python-dotenv
huggingface_hub
Pillow
google-genai